requirements.txt
argon2-cffi==18.1.0
boto3==1.5.20
botocore==1.8.34
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
Django==1.11.7
django-extensions==1.9.9
django-recaptcha2==1.0.3
django-redis==4.8.0
django-storages==1.6.5
docutils==0.14
enum34==1.1.6
futures==3.2.0
idna==2.6
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
pydot==1.2.4
pygraphviz==1.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
sorl-thumbnail==12.4.1
typing==3.6.2
urllib3==1.22
